Move file related stuff to their own file
