+++ lua/autorun/server/advdupe2_sv_init.lua
@@ -11,11 +11,13 @@ include "advdupe2/sv_clipboard.lua"
 include "advdupe2/sh_codec.lua"
 include "advdupe2/sh_netstream.lua"
 include "advdupe2/sv_misc.lua"
+include "advdupe2/sv_file.lua"
 
 AddCSLuaFile "autorun/client/advdupe2_cl_init.lua"
 AddCSLuaFile "advdupe2/file_browser.lua"
 AddCSLuaFile "advdupe2/sh_codec.lua"
 AddCSLuaFile "advdupe2/sh_netstream.lua"
+AddCSLuaFile "advdupe2/cl_file.lua"
 
 function AdvDupe2.Notify(ply,msg,typ, showsvr, dur)
 	umsg.Start("AdvDupe2Notify",ply)
